+++ CHANGELOG.md
@@ -22,6 +22,8 @@ We based the format on [Keep a Changelog][] and this project adheres to
 [Semantic Versioning][].
 
 ## [Unreleased][]
+### Added
+- Choose names transaction.
 
 ## [0.2.2][] - 2017-11-07
 ### Fixed